improved readability, cleaned source and added comments
