+++ src/look.js
@@ -1,41 +1,10 @@
-import Look, {createStyleSheet} from './core'
+import Look from './core'
+
 import State from './api/State'
+import StyleSheet from './api/StyleSheet'
 import Listener from './api/Listener'
 
-import alternativeValue from './plugins/alternativeValue'
-import statefulValue from './plugins/statefulValue'
-import customProperty from './plugins/customProperty'
-
-import {equal, unEqual, bigger, smaller, biggerThan, smallerThan} from './keys/condition'
-import {firstChild, lastChild, onlyChild, nthChild, nthLastChild} from './keys/pseudoClasses/childIndex'
-import {firstOfType, lastOfType, onlyOfType, nthOfType, nthLastOfType} from './keys/pseudoClasses/childIndex'
-import empty from './keys/pseudoClasses/empty'
-import extend from './keys/extend'
-
-const config = {
-	plugins : [customProperty, alternativeValue, statefulValue],
-	keys: {
-		//NOTE: Ordner matters! 
-		'>=': biggerThan,
-		'<=': smallerThan,
-		'!=': unEqual,
-		'>': bigger,
-		'<': smaller,
-		'=': equal,
-		extend: extend,
-		':empy': empty,
-		':first-child': firstChild,
-		':last-child': lastChild,
-		':only-child': onlyChild,
-		':nth-child': nthChild,
-		':nth-last-child': nthLastChild,
-		':first-of-type': firstOfType,
-		':last-of-type': lastOfType,
-		':only-of-type': onlyOfType,
-		':nth-of-type': nthOfType,
-		':nth-last-of-type': nthLastOfType
-	}
-}
+import config from './preconfig/native'
 
 export default (...args) => {
 	return Look(...args, config)
@@ -44,5 +13,5 @@ export default (...args) => {
 export {
 	Listener,
 	State,
-	createStyleSheet
+	StyleSheet
 }
