+++ src/plugins/prefixer.js
@@ -1,14 +1,20 @@
 import Prefixer from 'inline-style-prefixer'
-let userAgent = typeof navigator !== 'undefined' ? navigator.userAgent : undefined
 
 let prefixer = new Prefixer()
 
-export default (styles) => {
-  if (userAgent !== undefined) {
-    styles = prefixer.prefix(styles)
-    return styles
-  } else {
-    console.warn('Autoprefixing failed as there is no valid userAgent specified.')
-    console.warn('Use Config.setUserAgent to specify a custom userAgent for server-side rendering.')
+export default (styles, scopeArgs, config) => {
+
+  // replace userAgent if config provides alternative one
+  if (config.userAgent !== undefined) {
+    if (prefixer._userAgent !== config.userAgent) {
+      prefixer = new Prefixer(config.userAgent)
+    }
   }
+
+  // only prefix if userAgent is not undefined
+  if (prefixer._userAgent !== undefined) {
+    styles = prefixer.prefix(styles)
+  }
+
+  return styles
 }
